+++ BigFade.cpp
@@ -128,8 +128,8 @@ namespace Assets
 	uint8_t CBigFade::FindColor( const Color& color )
 	{
 		for ( size_t i = 0; i < k_uNumColors; i++ ) {
-			if ( std::memcmp( &color, m_Palette.GetColor( i ), sizeof( Color ) ) == 0 ) {
-				return (uint8_t)i;
+			if ( std::memcmp( &color, m_Palette.GetColor( static_cast<uint8_t>(i) ), sizeof( Color ) ) == 0 ) {
+				return static_cast<uint8_t>(i);
 			}
 		}
 		return m_Palette.GetColorKey( 0 );
